Adding login, signup, and signout
